feat: MediaData -> AttachmentData
MediaData was renamed to AttachmentData and moved to it's own file.
This was done as that model is re-used across other models as well,
and it did not belong in the channels file as such.
